+++ cypress/integration/MAT/WebInterface/508/508Test.spec.js
@@ -12,12 +12,12 @@ describe('508', () => {
   afterEach('Log Out', () => {
     //login.matLogout()
   })
-  it('508 test check for all', () => {
+  it.skip('508 test check for all', () => {
 
     cy.checkA11y()
 
   })
-  it('specify values in options', () => {
+  it.skip('specify values in options', () => {
     cy.checkA11y( null,{
       runOnly: {
         type: 'tag',
@@ -25,18 +25,18 @@ describe('508', () => {
       } //['wcag21aa', 'wcag2aa', 'best-practice', 'section508']
     })
   })
-  it('Has no detectable a11y violations on load (filtering to only include critical impact violations)', () => {
+  it.skip('Has no detectable a11y violations on load (filtering to only include critical impact violations)', () => {
     // Test on initial load, only report and assert for critical impact items
     cy.checkA11y(null, {
       includedImpacts: ['critical', 'serious']
     })
   })
-  it('Only logs a11y violations while allowing the test to pass', () => {
+  it.skip('Only logs a11y violations while allowing the test to pass', () => {
     // Do not fail the test when there are accessibility failures
     cy.checkA11y(null, null, null, {skipFailures: true})
 
   })
-  it('508 test check for one element or section', () => {
+  it.skip('508 test check for one element or section', () => {
     cy.checkA11y(loginUI.signInButton)
     cy.checkA11y('.auth-footer')
   })
