+++ package.json
@@ -33,8 +33,15 @@
         "dev:all:mat:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/**/*.spec.js' --browser chrome;",
         "dev:mat:smoke:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/SmokeTests/**/*.spec.js' --browser chrome;",
         "dev:mat:measurecomposer:measurepackager:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureComposer/MeasurePackager/**/*.spec.js' --browser chrome;",
+        "dev:mat:measurecomposer:cqlworkspace:codes:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureComposer/CQLWorkspace/Codes.spec.js' --browser chrome;",
+        "dev:mat:measurecomposer:cqlworkspace:libaryeditor:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureComposer/CQLWorkspace/MeasureComposerLibraryEditor.spec.js' --browser chrome;",
+        "dev:mat:measurecomposer:measuredetails:measurementperiod:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureComposer/MeasureDetails/MeasurementPeriod/**/*.spec.js' --browser chrome;",
+        "dev:mat:measurecomposer:populationbasis:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureComposer/PopulationBasis/**/*.spec.js' --browser chrome;",
         "dev:mat:cqllibrary:search:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/CQL_Library/Search.spec.js' --browser chrome;",
+        "dev:mat:cqllibrary:fhirlibraryconversion:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/CQL_Library/FHIRLibraryConversion/SuccessfulLibraryConversion.spec.js' --browser chrome;",
         "dev:mat:measurecomposer:cqlworkspace:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureComposer/CQLWorkspace/**/*.spec.js' --browser chrome;",
+        "dev:mat:measurelibary:fhirmeasureconversion:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureLibrary/FHIRMeasureConversion/**/*.spec.js' --browser chrome;",
+        "dev:mat:measurelibary:grids:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/MeasureLibrary/Grids/**/*.spec.js' --browser chrome;",
         "dev:all:bonnieFHIR:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=bonnieFHIRDev --spec 'cypress/integration/BonnieFHIR/**/*.spec.js' --browser chrome;",
         "test:all:bonnieFHIR:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=bonnieFHIRTest --spec 'cypress/integration/BonnieFHIR/**/*.spec.js' --browser chrome;",
         "dev:bonnieFHIR:execution:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/BonnieFHIR/WebInterface/Execution/**/*.spec.js' --browser chrome;",
@@ -75,11 +82,18 @@
         "dev:mat:cql:library:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:cql:library; npm run combine:reports; npm run generateOne:report",
         "dev:mat:cql:composer:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:cql:composer; npm run combine:reports; npm run generateOne:report",
         "dev:mat:cqllibrary:search:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:cqllibrary:search:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:cqllibrary:fhirlibraryconversion:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:cqllibrary:fhirlibraryconversion:tests; npm run combine:reports; npm run generateOne:report",
         "dev:mat:measure:composer:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measure:composer; npm run combine:reports; npm run generateOne:report",
         "dev:mat:measure:library:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measure:library; npm run combine:reports; npm run generateOne:report",
         "dev:mat:smoke:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:smoke:tests; npm run combine:reports; npm run generateOne:report",
         "dev:mat:measurecomposer:measurepackager:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurecomposer:measurepackager:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:measurecomposer:cqlworkspace:codes:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurecomposer:cqlworkspace:codes:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:measurecomposer:cqlworkspace:libaryeditor:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurecomposer:cqlworkspace:libaryeditor:tests; npm run combine:reports; npm run generateOne:report",
         "dev:mat:measurecomposer:cqlworkspace:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurecomposer:cqlworkspace:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:measurecomposer:measuredetails:measurementperiod:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurecomposer:measuredetails:measurementperiod:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:measurecomposer:populationbasis:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurecomposer:populationbasis:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:measurelibrary:fhirmeasureconversion:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurelibary:fhirmeasureconversion:tests; npm run combine:reports; npm run generateOne:report",
+        "dev:mat:measurelibrary:grids:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run dev:mat:measurelibary:grids:tests; npm run combine:reports; npm run generateOne:report",
         "test:all:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run test:all:tests; npm run combine:reports; npm run generateOne:report",
         "test:all:mat:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run test:all:mat:tests; npm run combine:reports; npm run generateOne:report",
         "test:mat:smoke:tests:report": "npm run delete:reports; npm run delete:mochawesome; npm run test:mat:smoke:tests; npm run combine:reports; npm run generateOne:report",
